+++ teneto/misc/corrcoef_matrix.py
@@ -1,5 +1,6 @@
 import numpy as np
-from scipy.stats import pearsonr, betai
+from scipy.stats import pearsonr
+from scipy.special import betainc
 
 
 def corrcoef_matrix(matrix):
@@ -9,7 +10,7 @@ def corrcoef_matrix(matrix):
     rf = r[np.triu_indices(r.shape[0], 1)]
     df = matrix.shape[1] - 2
     ts = rf * rf * (df / (1 - rf * rf))
-    pf = betai(0.5 * df, 0.5, df / (df + ts))
+    pf = betainc(0.5 * df, 0.5, df / (df + ts))
     p = np.zeros(shape=r.shape)
     p[np.triu_indices(p.shape[0], 1)] = pf
     p[np.tril_indices(p.shape[0], -1)] = pf